Static_Products

+++ lib/providers/web_cart_provider.dart
@@ -5,7 +5,7 @@ class WebCartItem {
   final String title;
   final String subtitle;
   final String image;
-  final int price;
+  final double price;
   int quantity;
 
   WebCartItem({
